分离App组件中的header成为Header组件
[Fix]
app中的header部分被拆分为Header组件。
